+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
 ## Upcoming
+* Enable/disable plugins in settings web UI
+* Enable some plugins by default
 * Allow SSO identity API for API token auth
 * Restore archive: different message filename per key
+
 
 
 ## v2026.55 - 2026-07-15